+++ audiostack/tests/content/test_folder.py
@@ -9,20 +9,11 @@ audiostack.api_key = os.environ["AUDIO_STACK_DEV_KEY"]  # type: ignore
 test_constants: dict = {}  #
 
 
-def test_list() -> None:
-    r = Folder.list(folder="")
-    print(r)
-    for fol in r.folders:
-        print(fol)
-    for fi in r.files:
-        print(fi)
-
-
 def test_create() -> None:
-    r = Folder.delete(folder="__PYTHON_TEST", delete_files=True)
-    r = Folder.create(name="__PYTHON_TEST")  # type: ignore
+    r = Folder.create(name="__PYTHON_TEST")
+    test_constants["folder_id"] = r.data["folderId"]
     print(r)
 
 
 def test_delete() -> None:
-    Folder.delete(folder="__PYTHON_TEST", delete_files=True)
+    Folder.delete(folder_id=test_constants["folder_id"])